Resource View page style revision and restructure
