--- plugin-hrm-form/src/components/resources/goo.js
@@ -1,285 +1,0 @@
-const goo = {
-  id: '2',
-  name: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services',
-  attributes: {
-    site: {
-      '507f1f77bcf86cd799439012': {
-        name: [
-          { info: null, value: 'Calgary Family Services Site', language: 'en' },
-          { info: null, value: 'Calgary Family Services Site FR', language: 'fr' },
-        ],
-        details: [
-          {
-            info: {
-              description: 'Calgary Family Services Site description',
-              alternateNames: 'Calgary Family Services Site - Main',
-            },
-            value: 'site/507f1f77bcf86cd799439012/details',
-            language: 'en',
-          },
-          {
-            info: {
-              description: 'Calgary Family Services Site description FR',
-              alternateNames: 'Calgary Family Services Site - Main FR',
-            },
-            value: 'site/507f1f77bcf86cd799439012/details',
-            language: 'fr',
-          },
-        ],
-        location: {
-          address1: [{ info: null, value: '123 Street Name', language: '' }],
-          address2: [{ info: null, value: 'Suite 200', language: '' }],
-          city: [{ info: null, value: 'Calgary', language: '' }],
-          county: [{ info: null, value: 'Rocky View County', language: '' }],
-          postalCode: [{ info: null, value: 'A1B 2C3', language: '' }],
-          province: [{ info: { name: 'Alberta' }, value: 'AB', language: '' }],
-          country: [{ info: null, value: 'Canada', language: '' }],
-        },
-        email: [{ info: null, value: 'communications@calgaryfamily.org', language: '' }],
-        operations: {
-          0: [
-            {
-              info: { day: 'Monday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Walk-in only' },
-              value: 'Monday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Lundi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Sans rendez-vous' },
-              value: 'Lundi',
-              language: 'fr',
-            },
-          ],
-          1: [
-            {
-              info: { day: 'Tuesday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Tuesday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Mardi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Mardi',
-              language: 'fr',
-            },
-          ],
-          2: [
-            {
-              info: { day: 'Wednesday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Wednesday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Mercredi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Mercredi',
-              language: 'fr',
-            },
-          ],
-          3: [
-            {
-              info: { day: 'Thursday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Thursday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Jeudi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Jeudi',
-              language: 'fr',
-            },
-          ],
-          4: [
-            {
-              info: { day: 'Friday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Friday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Vendredi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'By appointment only' },
-              value: 'Vendredi',
-              language: 'fr',
-            },
-          ],
-          5: [
-            {
-              info: { day: 'Saturday', hoursOfOperation: null, descriptionOfHours: null },
-              value: 'Saturday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Samedi', hoursOfOperation: null, descriptionOfHours: null },
-              value: 'Samedi',
-              language: 'fr',
-            },
-          ],
-          6: [
-            {
-              info: { day: 'Sunday', hoursOfOperation: null, descriptionOfHours: null },
-              value: 'Sunday',
-              language: 'en',
-            },
-            {
-              info: { day: 'Dimanche', hoursOfOperation: null, descriptionOfHours: null },
-              value: 'Dimanche',
-              language: 'fr',
-            },
-          ],
-        },
-        phone: {
-          tollFree: [{ info: null, value: '1800123456', language: '' }],
-          hotline: [{ info: null, value: '14289876543', language: '' }],
-          outOfAreaLine: [{ info: null, value: '14161234567', language: '' }],
-          afterHoursLine: [{ info: null, value: '14289876542', language: '' }],
-          businessLine: [{ info: null, value: '14289876540', language: '' }],
-          fax: [{ info: null, value: '14289876544', language: '' }],
-          tty: [{ info: null, value: '14289876545', language: '' }],
-        },
-        isActive: [{ info: null, value: 'true', language: '' }],
-      },
-    },
-    nameDetails: [
-      {
-        info: {
-          official: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services',
-          alternate: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services alternate',
-        },
-        value: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services',
-        language: 'en',
-      },
-      {
-        info: {
-          official: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services FR',
-          alternate: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services alternate FR',
-        },
-        value: 'Calgary Family Services - Child/Youth/Family Counselling at Calgary Family Services FR',
-        language: 'fr',
-      },
-    ],
-    operations: {
-      0: [
-        {
-          info: { day: 'Monday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours' },
-          value: 'Monday',
-          language: 'en',
-        },
-        {
-          info: { day: 'Lundi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours FR' },
-          value: 'Lundi',
-          language: 'fr',
-        },
-      ],
-      1: [
-        {
-          info: { day: 'Tuesday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours' },
-          value: 'Tuesday',
-          language: 'en',
-        },
-        {
-          info: { day: 'Mardi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours FR' },
-          value: 'Mardi',
-          language: 'fr',
-        },
-      ],
-      2: [
-        {
-          info: { day: 'Wednesday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours' },
-          value: 'Wednesday',
-          language: 'en',
-        },
-        {
-          info: { day: 'Mercredi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours FR' },
-          value: 'Mercredi',
-          language: 'fr',
-        },
-      ],
-      3: [
-        {
-          info: { day: 'Thursday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours' },
-          value: 'Thursday',
-          language: 'en',
-        },
-        {
-          info: { day: 'Jeudi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours FR' },
-          value: 'Jeudi',
-          language: 'fr',
-        },
-      ],
-      4: [
-        {
-          info: { day: 'Friday', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours' },
-          value: 'Friday',
-          language: 'en',
-        },
-        {
-          info: { day: 'Vendredi', hoursOfOperation: '12pm - 5pm', descriptionOfHours: 'Call for hours FR' },
-          value: 'Vendredi',
-          language: 'fr',
-        },
-      ],
-      5: [
-        {
-          info: { day: 'Saturday', hoursOfOperation: null, descriptionOfHours: null },
-          value: 'Saturday',
-          language: 'en',
-        },
-        { info: { day: 'Samedi', hoursOfOperation: null, descriptionOfHours: null }, value: 'Samedi', language: 'fr' },
-      ],
-      6: [
-        { info: { day: 'Sunday', hoursOfOperation: null, descriptionOfHours: null }, value: 'Sunday', language: 'en' },
-        {
-          info: { day: 'Dimanche', hoursOfOperation: null, descriptionOfHours: null },
-          value: 'Dimanche',
-          language: 'fr',
-        },
-      ],
-    },
-    mainContact: {
-      name: [{ info: null, value: 'First Last', language: '' }],
-      email: [{ info: null, value: 'communications@calgaryfamily.org', language: '' }],
-      title: [
-        { info: null, value: 'Administrative Director', language: 'en' },
-        { info: null, value: 'Administrative Director FR', language: 'fr' },
-      ],
-      phoneNumber: [{ info: null, value: '4161234567', language: '' }],
-      isPrivate: [{ info: null, value: 'false', language: '' }],
-    },
-    website: [{ info: null, value: 'https://www.calgaryfamily.org', language: '' }],
-
-    description: [
-      { info: { text: 'Description' }, value: 'description', language: 'en' },
-      { info: { text: 'Description FR' }, value: 'description', language: 'fr' },
-    ],
-
-    primaryLocationCity: [{ info: null, value: 'Calgary', language: '' }],
-    primaryLocationCounty: [{ info: null, value: 'Rocky View County', language: '' }],
-    primaryLocationPostalCode: [{ info: null, value: 'A1B 2C3', language: '' }],
-    primaryLocationPhone: [{ info: null, value: '4032699888', language: '' }],
-    primaryLocationIsPrivate: [{ info: null, value: 'false', language: '' }],
-
-    keywords: [{ info: null, value: 'education', language: '' }],
-    taxonomyCode: [{ info: null, value: 'HP1234-01', language: '' }],
-    eligibilityMinAge: [{ info: null, value: '10', language: '' }],
-    eligibilityMaxAge: [
-      { info: null, value: '20', language: '' },
-      { info: null, value: 'Free', language: '' },
-    ],
-    agency: { isLocationPrivate: [{ info: null, value: 'false', language: '' }] },
-    interpretationTranslationServicesAvailable: [{ info: null, value: 'true', language: '' }],
-    documentsRequired: [{ info: null, value: 'Health Card,Birth Certificate', language: '' }],
-    available247: [{ info: null, value: 'true', language: '' }],
-    isPrivate: [{ info: null, value: 'false', language: '' }],
-    sourceCreatedAt: [{ info: null, value: '2013-04-01 00:00:00', language: '' }],
-    sourceUpdatedAt: [{ info: null, value: '2018-06-22 10:21:19', language: '' }],
-    status: [{ info: null, value: 'Active', language: '' }],
-    primaryLocationProvince: [{ info: { name: 'Alberta' }, value: 'AB', language: '' }],
-    targetPopulation: { 0: [{ info: null, value: 'Open To All', language: '' }] },
-    howToAccessSupport: [{ info: null, value: 'Appointment Required', language: '' }],
-    howIsServiceOffered: [{ info: null, value: 'In-Person Support', language: '' }],
-    accessibility: [{ info: null, value: 'Wheelchair Accessible', language: '' }],
-    applicationProcess: [
-      {
-        info: null,
-        value:
-          'Application process unknown. A step in the right direction would be to call or email to connect and learn more.',
-        language: '',
-      },
-    ],
-  },
-};
