+++ plugin-hrm-form/src/styles/resources/index.ts
@@ -1,0 +1,27 @@
+/**
+ * Copyright (C) 2021-2023 Technology Matters
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see https://www.gnu.org/licenses/.
+ */
+
+import { styled } from '@twilio/flex-ui';
+
+import { Absolute } from '../HrmStyles';
+
+// ViewResource Page
+export const ResourceViewContainer = styled(Absolute)`
+  height: 100%;
+  width: 1280px;
+  background-color: #f6f6f6;
+`;
+ResourceViewContainer.displayName = 'ResourceViewContainer';
